fastapi-test/requirements.txt
# retinaface, insightface, fastapi 정상작동.

absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.1
astunparse==1.6.3
beautifulsoup4==4.14.2
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.1
debugpy==1.8.17
decorator==5.2.1
easydict==1.13
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.121.2
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
gdown==5.2.0
google-auth==2.43.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.14.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.0
insightface==0.7.3
ipykernel==6.31.0
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.12.0
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==1.26.4
oauthlib==3.3.1
onnx==1.19.1
onnxruntime==1.23.2
opencv-python==4.9.0.80
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prettytable==3.17.0
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
requests-oauthlib==2.0.0
retina-face @ file:///Users/yunha_hwa_ng/Documents/DRAWER/career/2511%20NIPA-Google%20team/retina-face-0.0.17.tar.gz#sha256=7532b136ed01fe9a8cba8dfbc5a046dd6fb1214b1a83e57f3210bd145a91cd73
rich==14.2.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.49.3
stringzilla==4.2.3
sympy==1.14.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tf_keras==2.16.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.14.2
zipp==3.23.0
